+++ src/compiler/util.ts
@@ -4,11 +4,14 @@ import {NodeContext} from "./error";
 import {printRaw} from "@unified-latex/unified-latex-util-print-raw";
 import {EXIT, visit} from "@unified-latex/unified-latex-util-visit";
 import {Pluggable} from "unified";
+import {getParser} from "@unified-latex/unified-latex-util-parse";
 
 
 export type RenderToHtml = (node: Node) => Promise<string>;
 export type RenderPlugin = (root: Node) => void;
 export type RendererBuilder = (plugins: Pluggable[]) => RenderToHtml;
+
+export type Parser = ReturnType<typeof getParser>;
 
 // Root folder where all the graphics will be deposited.
 export const graphicsRoot = "./public/g/";