+++ changelog.md
@@ -7,7 +7,9 @@
 - Reorganised the usage documentation to not remain on one page. 
 - Fixed a bug where enumerates require an additional line break after.
 - Fixed a bug where Firefox would not load XyPic properly due to not trusting a non-npm jsdelivr link. 
-- Fixed a bug where using hythens in the search bar crashes the website.
+- Fixed a bug where using hyphens in the search bar crashes the website.
+- Fixed a bug where conservative mode does not recognise custom environments.
+- Fixed a bug where custom environments do not accept its optional argument. 
 
 
 # v0.2.3, 2026-04-23